Consolidated routes files into api/items.js, removed app/items.js. Attempted JWT authorization failure debugging. Unsucessful.
