+++ .gitignore
@@ -5,3 +5,4 @@ public/bundle.js.map
 npm-debug.log
 yarn-error.log
 package-lock.json
+.env
