+++ server/api/items.js
@@ -4,6 +4,42 @@ const {
 } = require("../db");
 module.exports = router;
 
+async function requireAdmin(req, res, next) {
+  if (!req.user || !req.user.isAdmin) {
+    return res.sendStatus(403).send("You shall not pass!");
+  }
+}
+
+router.post("/", requireAdmin, async (req, res, next) => {
+  try {
+    const item = await Item.create(req.body);
+    res.json(item);
+  } catch (err) {
+    next(err);
+  }
+});
+
+router.put("/:itemId", async (req, res, next) => {
+  try {
+    const item = await Item.findByPk(req.params.itemId);
+    if (!product) return res.sendStatus(404);
+    await item.update(req.body);
+    res.json(item);
+  } catch (err) {
+    next(err);
+  }
+});
+
+router.delete("/:itemId", requireAdmin, async (req, res, next) => {
+  try {
+    const item = await Item.findByPk(req.params.itemId);
+    await item.destroy();
+    res.sendStatus(204);
+  } catch (err) {
+    next(err);
+  }
+});
+
 router.get("/items", async (req, res, next) => {
   try {
     const items = await Item.findAll();
